Add files via upload
This is an assignment work which describes the design and implementation of a system that integrates Internet of Things (IoT) and Machine Learning (ML) technologies to track and analyze heart rate (BPM), SPO2 and body temperature levels. The body temperature is associated with a number of the human heartbeat; little change in body temperature can significantly affect the performance of high cardiac health problems. The main purpose of the proposed system is to predict the occurrence of cardiovascular disease. The system employs a Raspberry Pi 4, a MAX30102 sensor for the detection of beats per minute (BPM) and SPO2 levels, an AMG8833 sensor for measuring body temperature and Thinkspeak for cloud storage and insights visualization. Using binary classification (classify instances into one of two possible categories or classes) improves the system's ability to reliably predict the occurrence of cardiac disease, allowing for proactive care and prevention methods.